+++ auth-app/src/main/java/org/gameontext/auth/Health.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package net.wasdev.gameon.auth;
+package org.gameontext.auth;
 
 import java.io.IOException;
 import java.util.logging.Level;
