+++ auth-app/src/main/java/org/gameontext/auth/facebook/FacebookAuth.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package net.wasdev.gameon.auth.facebook;
+package org.gameontext.auth.facebook;
 
 import java.io.IOException;
