+++ auth-app/src/main/java/org/gameontext/auth/github/GitHubAuth.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package net.wasdev.gameon.auth.github;
+package org.gameontext.auth.github;
 
 import java.io.IOException;
 import java.util.UUID;
