+++ auth-app/src/main/java/org/gameontext/auth/google/GoogleAuth.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package net.wasdev.gameon.auth.google;
+package org.gameontext.auth.google;
 
 import java.io.IOException;
 import java.util.Arrays;
