+++ auth-app/src/main/java/org/gameontext/auth/twitter/TwitterAuth.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package net.wasdev.gameon.auth.twitter;
+package org.gameontext.auth.twitter;
 
 import java.io.IOException;
