+++ auth-app/src/test/java/org/gameontext/auth/CertificateUtils.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  *******************************************************************************/
-package net.wasdev.gameon.player;
+package org.gameontext.auth;
 
 import static org.junit.Assert.assertNotNull;
